backend/requirements.txt
# ExamBuddy Backend - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
mangum==0.17.0  # AWS Lambda adapter for FastAPI

# AWS SDK
boto3==1.29.7
botocore==1.32.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation & Parsing
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# PDF Processing
pypdf==3.17.1
pdfplumber==0.10.3

# Report Generation
reportlab==4.0.7
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For FastAPI test client
moto[dynamodb,s3]==4.2.9  # AWS service mocking

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
